Get unit tests working; fix axis label staggering
This is really a wholistic copy of the unit test frame work from the
recent novus/nvd3.git/massive-merge branch by Robin Hu. See:

https://github.com/novus/nvd3/commits/massive-merge

In order to get all the tests working, we have back down to D3 v3.3.13
(see Robin's comment in the README.md of his merge), and fix a bug with
staggered labels. See:

https://github.com/novus/nvd3/commit/65134d2fc2ee65a1d9dcebc46e015c0d9da568a6

With the above we have passing unit tests so that we can begin to add
further tests as we make changes.